% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengene.R
\name{eigengene}
\alias{eigengene}
\title{Inference of differential co-expression using the eigengene approach}
\usage{
eigengene(
  x,
  y,
  type_x,
  mean_adjust = matrix(rep(1, length(x), ncol = 1)),
  var_adjust = matrix(rep(1, length(x), ncol = 1)),
  perm.its = 1000,
  return.stats = FALSE
)
}
\arguments{
\item{x}{The primary variable(s) of interest (n observations by p variables)}

\item{y}{A matrix of genes  in a gene set (n observations by k genes)}

\item{type_x}{Specify type of primary variable. "categorical" or "continuous.}

\item{mean_adjust}{A matrix of covariates to adjust for mean effects.}

\item{var_adjust}{A matrix of covariates to adjust for variance effects.}

\item{perm.its}{The number of permutations. The default is 1000.}

\item{return.stats}{If TRUE then the permutation statistics are returned. The default is FALSE.}
}
\value{
A list of object type "kdca" containing:
\item{call}{Function call.}
\item{pvalues}{An empirical p-value.}
\item{null.stats}{If return.stats is TRUE, then a vector of null statistics is returned.}
\item{obs.stats}{If return.stats is TRUE, then the observed test statistic is returned.}
}
\description{
Given a risk factor of interest, the \code{eigengene} function tests for
differential co-expression of a pathway using the top eigengene of the correlation matrix.
}
\examples{
# set seed
set.seed(123)

# Generate random study design and gene set
x <- matrix(rbinom(50, size = 1, prob = 0.5), ncol = 1)
y <- matrix(rnorm(50 * 10), ncol = 10)

eg.out <- eigengene(x, y, type_x = "categorical")
}
